backend/requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
django-cors-header==3.3.0
django-background-tasks==1.2.5